+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/ElementConfigurationSectionNew.test.tsx
@@ -18,7 +18,7 @@ import * as React from 'react';
 import { render, screen, waitFor } from 'wrappedTestingLibrary';
 import userEvent from '@testing-library/user-event';
 
-import ElementConfigurationContainer from './ElementConfigurationContainer';
+import ElementConfigurationSection from './ElementConfigurationSectionNew';
 
 describe('ElementConfigurationContainer', () => {
   it('should render the component with children', async () => {
